pyproject.toml
[project]
name = "chunkflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bm25s>=0.2.14",
    "paddleocr>=3.3.1",
    "paddlepaddle>=3.2.1",
    "pymupdf>=1.26.6",
    "pytest>=9.0.0",
    "python-docs>=0.1.0",
    "python-docx>=1.2.0",
    "rank-bm25>=0.2.2",
    "typer>=0.20.0",
]
